Implementando validações nas API's REST de Marca.

+++ app/Http/Controllers/MarcaController.php
@@ -4,9 +4,16 @@ namespace App\Http\Controllers;
 
 use App\Models\Marca;
 use Illuminate\Http\Request;
+use Ramsey\Uuid\Type\Integer;
 
 class MarcaController extends Controller
 {
+
+    public function __construct(Marca $marca)
+    {
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,8 +21,8 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marcas = Marca::all();
-        return $marcas;
+        $marcas = $this->marca->all();
+        return response()->json($marcas,200);
     }
 
     /**
@@ -36,19 +43,25 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca = Marca::create($request->all());
-        return $marca;
+
+        $request->validate($this->marca->rules(),$this->marca->feedback());
+        $marca = $this->marca->create($request->all());
+        return response()->json($marca,201);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
-        return $marca;
+        $marca = $this->marca->find($id);
+        if ($marca === null){
+            return response()->json(['erro' => 'Recurso pesquisado não existe'],404);
+        }
+        return response()->json($marca,200);
     }
 
     /**
@@ -66,24 +79,38 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request,$id)
     {
+
+        $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return response()->json(['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'],404);
+        }
+
+        $request->validate($marca->rules(),$marca->feedback());
         $marca->update($request->all());
-        return $marca;
+        return response()->json($marca,200);
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return response()->json(['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'],404);
+        }
+
         $marca->delete();
-        return ['msg'=> 'A marca foi removida com sucesso!'];
+        return response()->json(['msg'=> 'A marca foi removida com sucesso!'],200);
     }
 }
